requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
passlib
fastapi-versioning
email-validator
